Swap if branches for asserts in get/set surface
The if branches were effectively dead code, as the original
InstrumentWidget (and InstrumentDisplay) always have a valid ptr for
both members.

Because of this we can replace the if branches with the assumption the
pointers will always be true, as they have been thus far. We've put in
an assert in-case this no longer holds true in the future too.