Remove logic from presenter and use new model logic
Re #31929
Removes the logic to decide whether to tick save options (which was
broken) out of presenter and use new logic in model.

Whilst the tests removed consider if we are in memory mode, we
completely ignore this now. Since we want to update the options
regardless and the model will ultimately ignore the options regardless.

This massively simplifies the logic for something minor visually.